Production Report By Broker: Update to match proc

+++ Codes.Service/ViewModels/Reports/ProductionResultSummaryViewModel.cs
@@ -28,7 +28,7 @@ namespace ClientPortal.Models
         public decimal TotalYouPayPrice { get; set; }
 
         [DisplayFormat(DataFormatString = "{0:c}")]
-        public double TotalMemberSavings { get; set; }
+        public decimal TotalMemberSavings { get; set; }
 
         [DisplayFormat(DataFormatString = "{0:c}")]
         public decimal TotalCommissionEarned { get; set; }
@@ -56,7 +56,7 @@ namespace ClientPortal.Models
         public decimal TotalYouPayPrice { get; set; }
 
         [DisplayFormat(DataFormatString = "{0:c}")]
-        public double TotalMemberSavings { get; set; }
+        public decimal TotalMemberSavings { get; set; }
 
         [DisplayFormat(DataFormatString = "{0:c}")]
         public decimal TotalCommissionEarned { get; set; }
@@ -73,7 +73,7 @@ namespace ClientPortal.Models
         public decimal YouPayPrice { get; set; }
 
         [DisplayFormat(DataFormatString = "{0:c}")]
-        public double MemberSavings { get; set; }
+        public decimal MemberSavings { get; set; }
 
         [DisplayFormat(DataFormatString = "{0:c}")]
         public decimal CommissionEarned { get; set; }
